Save Copy of File Fetched from Main server
New function "saveToFile"
now the server functions as a proxy server regarding files named "json" and "anything".
the server saves a permanent copy of the file's contents in the "Files" directory as a .txt file.
in future requests regarding "json" or "anything" get requests, they are also treated as "json.txt" or "anything.txt" when looking for a matching file name.